Create webtoken.sol
This is a Solidity code for creating an ERC20 token using OpenZeppelin contracts. The code defines a contract called "W3BVIII" which inherits the ERC20 standard. The contract constructor takes an amount parameter, which is used to mint the tokens to the contract address. The token price is set to 2 ether. The "buyToken" function allows users to buy tokens by sending ether to the contract, and the "withdraw" function allows the owner to withdraw ether from the contract. The contract also has "receive" and "fallback" functions for receiving and handling incoming payments, respectively.